feat: Add RowColorCatalog component for custom row colors
The code changes include adding a new component called RowColorCatalog, which defines the colors for different levels of table headers and rows. This component is used to customize the colors of the table rows based on their depth. The RowColorCatalog component provides a catalog of color values that can be easily referenced and used throughout the application.